requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-core==1.35.0
azure-identity==1.24.0
azure-keyvault-secrets==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
distro==1.9.0
fastapi==0.116.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
msal==1.33.0
msal-extensions==1.3.1
openai==1.99.6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.4
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.33.0
wheel==0.45.1
